+++ index.php
@@ -46,7 +46,9 @@ if($pv_meta['app_ver'] < trim(file_get_contents(__ROOT__.'/VERSION.md'))){
   <meta name="description" content="<?php echo $product.' - '.$ver;?>">
   <meta name="author" content="JBPARFUM">
   <title><?php echo $product;?> - Dashboard</title>
-  
+  <link rel="icon" type="image/png" sizes="32x32" href="img/favicon-32x32.png">
+  <link rel="icon" type="image/png" sizes="16x16" href="img/favicon-16x16.png">
+
   <link href="css/fontawesome-free/css/all.min.css" rel="stylesheet">
   <link href="css/sb-admin-2.css" rel="stylesheet">
   <link href="css/bootstrap-select.min.css" rel="stylesheet">