+++ pages/viewFormula.php
@@ -247,13 +247,13 @@ $('.replaceIngredient').editable({
                     <tr>
                       <?php if($settings['grp_formula'] == '1'){ echo '<th class="noexport"></th>'; } ?>
                       <th width="22%">Ingredient</th>
-                      <th width="10%">CAS #</th>
-                      <th width="10%">Purity %</th>
-                      <th width="10%">Dilutant</th>
-                      <th width="10%">Quantity (ml)</th>
-                      <th width="10%">Concentration*</th>
-                      <th width="10%">Cost</th>
-                      <th width="10%">Properties</th>
+                      <th width="5%">CAS #</th>
+                      <th width="5%">Purity %</th>
+                      <th width="5%">Dilutant</th>
+                      <th width="5%">Quantity (ml)</th>
+                      <th width="5%">Concentration*</th>
+                      <th width="5%">Cost</th>
+                      <th width="5%">Properties</th>
                       <th class="noexport" width="15%">Actions</th>
                     </tr>
                   </thead>
@@ -287,10 +287,12 @@ $('.replaceIngredient').editable({
 								echo '<td class="noexport">'.$ing_q['profile'].'</td>';
 							}
 						}
-                      echo '<td align="center" class="'.$ing_q['profile'].'" id="ingredient"><a href="pages/mgmIngredient.php?id='.$formula['ingredient'].'" class="popup-link">'.$ingName.'</a> '.checkIng($formula['ingredient'],$settings['defCatClass'],$conn).'</td>
-					  <td align="center">'.$ing_q['cas'].'</td>
-                      <td data-name="concentration" class="concentration" data-type="text" align="center" data-pk="'.$formula['ingredient'].'">'.$formula['concentration'].'</td>';
-					  if($formula['concentration'] == '100'){
+						?>
+                      <td align="center" class="<?php if($settings['grp_formula'] == '0'){echo $ing_q['profile'];}?>" id="ingredient"><a href="pages/mgmIngredient.php?id=<?php echo $formula['ingredient'];?>" class="popup-link"><?php echo $ingName;?></a> <?php echo checkIng($formula['ingredient'],$settings['defCatClass'],$conn);?></td>
+                      <td align="center"><?php echo $ing_q['cas'];?></td>
+                      <td data-name="concentration" class="concentration" data-type="text" align="center" data-pk="<?php echo $formula['ingredient'];?>"><?php echo $formula['concentration'];?></td>
+					  <?php
+                      if($formula['concentration'] == '100'){
 						  echo '<td align="center">None</td>';
 					  }else{
 						  echo '<td data-name="dilutant" class="dilutant" data-type="select" align="center" data-pk="'.$formula['ingredient'].'">'.$formula['dilutant'].'</td>';