+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 ### Version 2.4
+- ADD: PubChem - exclude Mixtures and Blends
+- ADD: Default purity value '100' in ingredient management
 
 ### Version 2.3
 - ADD: Add a new dashboard