+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 # CHANGELOG
+### Version 2.4
+- ADD: PubChem - exclude Mixtures and Blends
+- ADD: Default purity value '100' in ingredient management
+- ADD: Favicon added!
+- CHG: Add new formula page now opens as a modal window
+- CHG: Allergen has been renamed to 'Declare'
+- CHG: Remove ingredient profile for grouped view in formula
+- IMP: Increase the size of the pop-up window when showing ingredient data
+- FIX: Minor Bug fixes and cosmetic changes
+- FIX: Dashboard pie graphs size
+- FIX: Docker helper script
+- FIX: Self update procedure
+
 ### Version 2.3
 - ADD: Add a new dashboard
 - ADD: Option to generate related documents to sell a formula